Add files via upload
This is a simple calculator with Python the inspiration for this script came from Lisa, T (2021) How To Make a Calculator Program in Python 3 (Version 2) [Computer Program], https://www.digitalocean.com/community/tutorials/how-to-make-a-calculator-program-in-python-3. With some changes to the script to make my own work. for starters as soon as the script is run the calculator takes in input from the user on what mathematical operation the user wants to run Lisa provided addition, subtraction, multiplication, and division and I added the power operation and modulo operation and also rewrote the program with a loop statement so that I can execute my code repeatedly. I also added an exit operation so that if you wanted to exit out of the calculator you could do so without error and a clear operation so that you can go back if the user inputted the wrong operation. The calculator takes in the operation that you want to use and then askes for the first number that you want to use and the second number that you want to use if you do anything that is not a number the program will prompt you to enter a valid number if you input numbers then it should show you the whole problem that it calculated and the answer (5 + 5 = 10) after that the program loops and asks for another operation I you do not want to use the program simply typing "exit" will make you exit the calculator.